my_rag_project/requirements.txt
# --- Core Dependencies ---
fastapi
uvicorn[standard]
streamlit
sentence-transformers


# --- RAG / Vector Store ---
faiss-cpu
langchain
langchain-community
chromadb

# --- LLM Providers ---
google-generativeai
langchain-google-genai
langchain-groq

# --- Embeddings / Text Processing ---
sentence-transformers
pypdf

python-dotenv

# --- Optional Utilities ---
tqdm
numpy
